Renamed authors ctrl to authors-public
no issue

- consistency
