+++ core/server/api/v2/authors-public.js
